Little fixes and combined tests to one module
